Use a single generic Envelope type for all structs
Each former struct corresponding to a Gtk object T is now an Envelope<T>.
While "bare impls" are not allowed we still need dummy structs to put constructors in. In the future the constructors would live in the impl Envelope<T>.
The traits don't have to keep track of object hierarchy, it's managed in a single place with cast_impl macros.
Make refcounting more correct.